move roleLabelTag
