+++ client/src/components/Domain/Shared/components/RoleLabelTag/types.ts
@@ -1,0 +1,8 @@
+import { LabelColor } from 'components/Core/DataDisplay/LabelTag/types'
+
+export enum Roles {
+    coordinator = 'Coördinator',
+    mentor = 'Begeleider',
+    volunteer = 'Vrijwilliger',
+    coworker = 'Medewerker',
+}
